requirements.txt
asgiref==3.3.4
Django==3.2.1
django-environ==0.4.5
django-filter==2.4.0
django-password-validators==1.6.0
djangorestframework==3.12.4
PyMySQL==1.0.2
pytz==2021.1
sqlparse==0.4.1
